requirements.txt
aiohttp==3.8.4
aiohttp_socks==0.8.0
comtypes==1.1.14
legofy==1.0.0
numpy==1.24.2
opencv_python==4.7.0.72
Pillow==9.5.0
python_telegram_bot==20.1
pywin32==305
PyYAML==6.0
requests==2.28.2
scikit_learn==1.2.2
telegram==0.0.1
Wand==0.6.11
zalgo_text==0.6
